requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
gitdb==4.0.12
gitpython==3.1.45
groq==0.31.0
gtts==2.5.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
langchain-core==0.3.74
langchain-ollama==0.3.6
langsmith==0.4.14
markupsafe==3.0.2
matplotlib-inline==0.1.7
mysql-connector-python==9.4.0
narwhals==2.1.2
nest-asyncio==1.6.0
numpy==2.2.6
ollama==0.5.3
orjson==3.11.2
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
postgrest==1.1.1
prompt-toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.2
realtime==2.7.0
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
speechrecognition==3.14.3
srt==3.5.3
stack-data==0.6.3
storage3==0.12.1
streamlit==1.48.1
strenum==0.4.15
supabase==2.18.1
supabase-auth==2.12.3
supabase-functions==0.10.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
vosk==0.3.44
wcwidth==0.2.13
websockets==15.0.1
youtube-transcript-api==1.2.2
zstandard==0.24.0
pdfplumber
fpdf2